Created Weapon Model Builder and Started Bullets
Created the behavior needed so that I can import and assign new weapon models to the scriptable object. These models will then be used to create the full weapon model accordingly to the type.

Improvement may be made to above function as scalability can become an issue with weapon building.

Started to create data storage for bullet models//game objects.